requirements.txt
pyqrcode==1.2.1
pypng==0.0.20